backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0
pydantic>=2.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0